DO-NOT-MERGE: libgccjit experiment (work-in-progress)
Inspired by Vladimir's talk at LPC, this is a proof-of-concept of
populating a libgccjit context from a MIR module, and then
compiling it with libgccjit.

This could allow for libgccjit to be used as a tier 2 JIT compiler
in conjunction with MIR for tier 1.

Additionally, Vladimir's talk identified a weakness of libgccjit:
the difficult of creating the environment through the API.  With
this approach, C could be compiled to MIR, and then injected into
libgccjit.

Right now it's a very rough proof-of-concept that only implements
a few instructions and modes.  In particular it doesn't yet support
forward references.  But it does support compiling trivial binary
mir files via libgccjit.

Given that you don't want external deps, I'm not sure where this
should ultimately live.  Also, I used C++ to simplify the
implementation, but could recode it in C if that's a deal-breaker.